Register range tests in the suite, fold CORS coverage into them
The realm-server test entrypoint loads an explicit file list, and neither
http-range-test nor range-request-test was on it, so none of the range or
CORS coverage was actually running in CI. Both are registered now.

The two CORS assertions move from their own module into range-request-test,
which stands up a byte-identical realm fixture — picking up its asserted
sample upload on the way. The exposure test now also pins
Access-Control-Allow-Origin, without which no expose list is consulted, and
the create-response comment states the real consumer of the exposed range
headers: the CORS-filtered Response the auth service worker hands the media
element via respondWith.
