+++ packages/realm-server/tests/index.ts
@@ -244,6 +244,8 @@ const ALL_TEST_FILES: string[] = [
   './file-watcher-events-test',
   './full-index-on-startup-test',
   './full-reindex-test',
+  './http-range-test',
+  './range-request-test',
   './http2-keepalive-test',
   './indexing-test',
   './lazy-mount-test',
