+++ packages/runtime-common/create-response.ts
@@ -28,9 +28,13 @@ export function createResponse({
       // which is why `Retry-After` appears below: readiness pairs it with
       // `X-Boxel-Not-Ready` on a 503, and a cross-origin caller that could read
       // the stage but not the retry hint would only have half the answer.
-      // Content-Range/Accept-Ranges let a cross-origin caller reason about a
-      // 206 byte-range response (Content-Length is safelisted, but is listed
-      // for symmetry with the range pair).
+      // Content-Range/Accept-Ranges must be exposed because the host's auth
+      // service worker hands the CORS-filtered Response from its own fetch()
+      // straight to the media element via respondWith. A header absent from
+      // this list is pruned from that filtered Response, so it is invisible
+      // to the player's loading stack, not just to app JS. Content-Length is
+      // CORS-safelisted and survives regardless; it is named alongside the
+      // pair so the range trio reads as one unit.
       'Access-Control-Expose-Headers':
         'X-Boxel-Realm-Url,X-Boxel-Realm-Public-Readable,X-Boxel-Realm-Archived,X-Boxel-Canonical-Path,X-Boxel-Not-Ready,Authorization,Cache-Control,ETag,Retry-After,Content-Range,Accept-Ranges,Content-Length',
     },
